Fix test imports and move testGoto
