+++ test/ast.d/function-statement.ts
@@ -1,7 +1,7 @@
 import * as _ from 'tap';
 import * as AST from 'src/ast.js';
 import * as Stakr from 'src/stakr.js';
-import { AssembleArg, ExecuteArg } from 'src/types';
+import { AssembleArg, ExecuteArg } from 'src/types.d';
 
 void _.test('FunctionStatement', (_) => {
 	void _.test('name', (_) => {
