+++ test/ast.d/label.ts
@@ -1,7 +1,7 @@
 import * as _ from 'tap';
 import * as AST from 'src/ast.js';
 import * as Stakr from 'src/stakr.js';
-import { AssembleArg } from 'src/types';
+import { AssembleArg } from 'src/types.d';
 
 void _.test('Label', (_) => {
 	void _.test('name', (_) => {
