+++ test/commands.d/goto.ts
@@ -2,27 +2,8 @@ import * as _ from 'tap';
 import * as Stakr from 'src/stakr.js';
 import url from 'url';
 import { goto_ } from 'src/commands.js';
-import { ExecuteArg } from 'src/types';
-
-export function testGoto (
-	_: Tap.Test,
-	command: () => void,
-	arg: ExecuteArg,
-) {
-	const { context } = arg;
-	context.stack.length = 0;
-	arg.offset = 0;
-	_.throws(command, 'expected to throw if stack is empty');
-	context.push('abc');
-	_.throws(command, 'expected to throw if poped value is a string');
-	context.push(true);
-	_.throws(command, 'expected to throw if poped value is not a number');
-	context.stack.length = 0;
-	context.push(123);
-	command();
-	_.equal(arg.offset, 123, 'expected to jump to given offset');
-	_.strictSame(context.stack, [], 'expected to pop from the stack');
-}
+import { ExecuteArg } from 'src/types.d';
+import { testGoto } from '../util/goto.js';
 
 if (process.argv[1] === url.fileURLToPath(import.meta.url)) {
 	const context = new Stakr.ExecutionContext();
