--- test/main.ts
@@ -1,3 +1,0 @@
-import * as _ from 'tap';
-
-_.plan(0);
